Generate bundle.js to include hash so it can be cached on production
The generated 'bundle-[hash].js' can be effectively cached on production
for long-term.

Also, make sure that the hot-reload functionality should work regardless
of this change which affects a production build.

Update the list of commands to include the link to
clicks.vishaltelangre.com.

Add '_blank' target attribute to certain links to open in a new tab.

Update README.md with the instructions on how to generate a production
build.